+++ core/service/src/test/java/com/loopers/core/service/payment/component/FailedPaymentStrategyIntegrationTest.java
@@ -200,196 +200,4 @@ class FailedPaymentStrategyIntegrationTest extends IntegrationTest {
                     .isEqualTo(expectedReason);
         }
     }
-
-    @Nested
-    @DisplayName("동시성 제어 - 동시 결제 실패 시 재고 복구 검증")
-    class ConcurrencyControl {
-
-        @Test
-        @DisplayName("같은 상품에 대해 동시에 여러 결제가 실패할 때 재고 복구 일관성을 보장한다")
-        void shouldMaintainStockConsistencyUnderConcurrentFailedPayments() throws InterruptedException {
-            // Given
-            long initialStock = testProduct.getStock().value();
-            int concurrentFailedPaymentCount = 5;
-            int stockPerPayment = 10;
-
-            List<Payment> concurrentPayments = new ArrayList<>();
-
-            for (int i = 0; i < concurrentFailedPaymentCount; i++) {
-                Order order = orderRepository.save(
-                        OrderFixture.createWith(testUser.getId())
-                );
-                orderItemRepository.save(
-                        OrderItemFixture.createWith(order.getId(), testProduct.getId(), stockPerPayment)
-                );
-
-                concurrentPayments.add(
-                        paymentRepository.save(PaymentFixture.createWith(order.getOrderKey(), testUser.getId()))
-                );
-            }
-
-            Product decreasedProduct = testProduct.decreaseStock(new Quantity((long) (concurrentFailedPaymentCount * stockPerPayment)));
-            productRepository.save(decreasedProduct);
-
-            List<PgPayment> pgPayments = new ArrayList<>();
-            for (int i = 0; i < concurrentFailedPaymentCount; i++) {
-                pgPayments.add(
-                        PgPaymentFixture.createWith(concurrentPayments.get(i).getId())
-                );
-            }
-
-            // When
-            ConcurrencyTestUtil.executeInParallelWithoutResult(
-                    concurrentFailedPaymentCount,
-                    index -> {
-                        PgPayment pgPayment = pgPayments.get(index);
-                        FailedReason failedReason = new FailedReason("결제 게이트웨이 오류 #" + index);
-                        failedPaymentStrategy.pay(pgPayment, failedReason);
-                    }
-            );
-
-            // Then
-            Product finalProduct = productRepository.findById(testProduct.getId()).orElseThrow();
-            assertThat(finalProduct.getStock().value())
-                    .as("동시 결제 실패로 인한 재고 복구가 정확하게 이루어져야 함")
-                    .isEqualTo(initialStock);
-        }
-
-        @Test
-        @DisplayName("10개의 동시 결제 실패에서 재고 복구 일관성을 보장한다")
-        void shouldMaintainStockConsistencyWith10ConcurrentFailedPayments() throws InterruptedException {
-            // Given
-            long initialStock = testProduct.getStock().value();
-            int concurrentFailedPaymentCount = 10;
-            int stockPerPayment = 5;
-
-            ArrayList<Payment> concurrentPayments = new ArrayList<>();
-
-            for (int i = 0; i < concurrentFailedPaymentCount; i++) {
-                Order order = orderRepository.save(
-                        OrderFixture.createWith(testUser.getId())
-                );
-                orderItemRepository.save(
-                        OrderItemFixture.createWith(order.getId(), testProduct.getId(), stockPerPayment)
-                );
-
-                concurrentPayments.add(
-                        paymentRepository.save(PaymentFixture.createWith(order.getOrderKey(), testUser.getId()))
-                );
-            }
-
-            Product decreasedProduct = testProduct.decreaseStock(new Quantity((long) (concurrentFailedPaymentCount * stockPerPayment)));
-            productRepository.save(decreasedProduct);
-
-            List<PgPayment> pgPayments = new ArrayList<>();
-            for (int i = 0; i < concurrentFailedPaymentCount; i++) {
-                pgPayments.add(
-                        PgPaymentFixture.createWith(concurrentPayments.get(i).getId())
-                );
-            }
-
-            // When
-            ConcurrencyTestUtil.executeInParallelWithoutResult(
-                    concurrentFailedPaymentCount,
-                    index -> {
-                        PgPayment pgPayment = pgPayments.get(index);
-                        FailedReason failedReason = new FailedReason("결제 실패: " + index);
-                        failedPaymentStrategy.pay(pgPayment, failedReason);
-                    }
-            );
-
-            // Then
-            Product finalProduct = productRepository.findById(testProduct.getId()).orElseThrow();
-            assertThat(finalProduct.getStock().value())
-                    .as("10개의 동시 결제 실패로 인한 재고 복구 일관성을 보장")
-                    .isEqualTo(initialStock);
-        }
-
-        @Test
-        @DisplayName("서로 다른 상품에 대한 동시 결제 실패에서 각 상품의 재고가 독립적으로 복구된다")
-        void shouldRestoreStockIndependentlyForDifferentProducts() throws InterruptedException {
-            // Given
-            Product product1 = testProduct;
-            Product product2 = productRepository.save(
-                    ProductFixture.createWith(
-                            new BrandId("2"),
-                            new ProductName("테스트 상품 2"),
-                            new ProductPrice(new BigDecimal("15000")),
-                            new ProductStock(1000L)
-                    )
-            );
-            Product product3 = productRepository.save(
-                    ProductFixture.createWith(
-                            new BrandId("3"),
-                            new ProductName("테스트 상품 3"),
-                            new ProductPrice(new BigDecimal("20000")),
-                            new ProductStock(1000L)
-                    )
-            );
-
-            long product1InitialStock = product1.getStock().value();
-            long product2InitialStock = product2.getStock().value();
-            long product3InitialStock = product3.getStock().value();
-
-            int concurrentFailedPaymentCount = 9; // 3개 상품 × 3개 결제씩
-            List<Payment> concurrentPayments = new ArrayList<>();
-
-            for (int i = 0; i < concurrentFailedPaymentCount; i++) {
-                Product targetProduct = switch (i / 3) {
-                    case 0 -> product1;
-                    case 1 -> product2;
-                    default -> product3;
-                };
-
-                Order order = orderRepository.save(
-                        OrderFixture.createWith(testUser.getId())
-                );
-                orderItemRepository.save(
-                        OrderItemFixture.createWith(order.getId(), targetProduct.getId(), 10L)
-                );
-
-                concurrentPayments.add(
-                        paymentRepository.save(PaymentFixture.createWith(order.getOrderKey(), testUser.getId()))
-                );
-            }
-
-            productRepository.save(product1.decreaseStock(new Quantity(30L)));
-            productRepository.save(product2.decreaseStock(new Quantity(30L)));
-            productRepository.save(product3.decreaseStock(new Quantity(30L)));
-
-            List<PgPayment> pgPayments = new ArrayList<>();
-            for (int i = 0; i < concurrentFailedPaymentCount; i++) {
-                pgPayments.add(
-                        PgPaymentFixture.createWith(concurrentPayments.get(i).getId())
-                );
-            }
-
-            // When
-            ConcurrencyTestUtil.executeInParallelWithoutResult(
-                    concurrentFailedPaymentCount,
-                    index -> {
-                        PgPayment pgPayment = pgPayments.get(index);
-                        FailedReason failedReason = new FailedReason("결제 실패 #" + index);
-                        failedPaymentStrategy.pay(pgPayment, failedReason);
-                    }
-            );
-
-            // Then
-            Product finalProduct1 = productRepository.findById(product1.getId()).orElseThrow();
-            Product finalProduct2 = productRepository.findById(product2.getId()).orElseThrow();
-            Product finalProduct3 = productRepository.findById(product3.getId()).orElseThrow();
-
-            assertSoftly(softly -> {
-                softly.assertThat(finalProduct1.getStock().value())
-                        .as("상품1 재고 복구")
-                        .isEqualTo(product1InitialStock);
-                softly.assertThat(finalProduct2.getStock().value())
-                        .as("상품2 재고 복구")
-                        .isEqualTo(product2InitialStock);
-                softly.assertThat(finalProduct3.getStock().value())
-                        .as("상품3 재고 복구")
-                        .isEqualTo(product3InitialStock);
-            });
-        }
-    }
 }
